Revert "Drop cpufrequtils as useful: Power management better left to kernel since Wheezy."
This reverts commit 45947aefc6afb21e968411c21191dc41aaf1acfc.

cpufreq-utils are very important for debugging and bring-up of embedded devices.

cpufreq-info can be used to check with operation points are used how often.

And with cpufreq-set we can force the cpu to a certain frequency to measure power consumption,
temperature etc.

The real power/frequency management is still left to the kernel if the kernel has
appropriate CONFIG and loads the correct govenors.