Letux-ssh-root: add a warning to /etc/motd that there is root access over ssh without password
During early stage development and especially low level debugging of kernels it
is in a firewalled environment much easier to simply ssh root@device or scp root@device
from debugging scripts than typing passwords or installing root certificates on the
device under test. Therefore we have introduced to install this package to open the
big security hole because it increases the convenience. Deinstallation is simply
an apt-get remove for this package

Signed-off-by: H. Nikolaus Schaller <hns@goldelico.com>